+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 # 2.0.0
 
     Added more flexibility around Filter Component
+
+# 2.1.0
+
+    Updated Dropdown component to allow search realtime 
+    Made option data components to get id for options